Use ResponseModel in patient_register api

+++ src/main/kotlin/com/icare/repository/UserRepositoryImpl.kt
@@ -1,8 +1,9 @@
 package com.icare.repository
 
-import com.google.firebase.auth.FirebaseAuth
 import com.icare.model.PatientModel
+import com.icare.model.ResponseModel
 import com.icare.model.Users
+import com.icare.utils.*
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.jdbc.core.JdbcTemplate
 import org.springframework.stereotype.Repository
@@ -14,37 +15,41 @@ class UserRepositoryImpl : UserRepository {
     @Autowired
     lateinit var iCareJdbcTemplate: JdbcTemplate
 
-    override fun registerPatient(patient: PatientModel): Boolean {
+    override fun registerPatient(patient: PatientModel): ResponseModel {
+        if (getUid(patient.token) == null) {
+            return ResponseModel(INVALID_TOKEN, null)
+        } else {
             val deleteUser = """
                 delete from Users where UserID = '${getUid(patient.token)}';
             """.trimIndent()
-        val sqlInsert = """ 
+            val sqlInsert = """ 
             INSERT into Patient([Patient_ID],[ChronicDiseases],[CurrentMedications],[Allergies],[PastSurgeries],[Weight])
         values ('${getUid(patient.token)}','${patient.chronicDiseases}','${patient.currentMedications}','${patient.allergies}','${patient.pastSurgeries}','${patient.weight}');""".trimIndent()
-        try {
-            if (insertUser(
-                    Users(
-                        getUid(patient.token),
-                        patient.roleID,
-                        patient.fName,
-                        patient.lName,
-                        patient.email,
-                        patient.birthDate,
-                        patient.gender,
-                        patient.isActive
+            try {
+                if (insertUser(
+                        Users(
+                            getUid(patient.token)!!,
+                            patient.roleID,
+                            patient.fName,
+                            patient.lName,
+                            patient.email,
+                            patient.birthDate,
+                            patient.gender,
+                            patient.isActive
+                        )
                     )
-                )
-            ) {
-                iCareJdbcTemplate.update(sqlInsert)
-                return true
-            }else{
-                return false
+                ) {
+                    iCareJdbcTemplate.update(sqlInsert)
+                    return ResponseModel(OK, null)
+                } else {
+                    return ResponseModel(DUPLICATE_USER, null)
+                }
+            } catch (e: Exception) {
+                println(e.stackTrace)
+                println(e.message)
+                iCareJdbcTemplate.update(deleteUser)
+                return ResponseModel(FAILED, null)
             }
-        } catch (e: Exception) {
-            println(e.stackTrace)
-            println(e.message)
-            iCareJdbcTemplate.update(deleteUser)
-            return false
         }
     }
 
@@ -69,12 +74,6 @@ class UserRepositoryImpl : UserRepository {
             println(e.message)
             return false
         }
-    }
-
-    private fun getUid(token: String): String {
-        val auth: FirebaseAuth = FirebaseAuth.getInstance()
-        val decodedToken = auth.verifyIdToken(token);
-        return decodedToken.uid;
     }
